pyproject.toml
[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 100
multi_line_output = 3
known_first_party = ["config"]

[tool.djlint]
profile = "django"
ignore = "H031"

[tool.poetry]
name = "HOT Wagtail CMS"
version = "0.1.0"
description = "Wagtail CMS for Humanitarian OpenStreetMap Team"
authors = ["Kat Olaguer <katherine.olaguer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
wagtail = "^6.2.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
docker = "^6.1.3"
gunicorn = "^21.2.0"
django-compressor = "^4.4"
django-browser-reload = "^1.12.1"
wagtail-localize = "^1.7"
wagtailmenus = "^4.0"
wagtail-modeladmin = "^2.0.0"
dj-database-url = "^2.2.0"
whitenoise = "^6.6.0"
django-storages = "^1.14.3"
boto3 = "^1.34.129"
wagtailgeowidget = "^8.1.1"
wand = "^0.6.13"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pytest-django = "^4.7.0"
pytest-timeout = "^2.2.0"
pytest-factoryboy = "^2.6.0"
django-browser-reload = "^1.12.1"
coverage = "^7.4.0"
django-pattern-library = "^1.2.0"
